Add TTML profiles
Pass through the profile features to check they are all the ones we want, and have the correct dispositions. Many removed, to do with visual presentation.

Added a formal TTML2 profile document for the Content Profile and the Processor Profile, because there's no way, currently, to express a mix of "optional" and "permitted" when inferring a Processor Profile from a Content Profile.

Defined some extension features to express formally some normative requirements, like the presence of `xml:lang` and `ttp:contentProfiles` on the root element.

Specify normatively the permission to use features from other profiles and add a note about IMSC as an example.